// web/src/types/generated/types.ts
/*
 Generated by typeshare 1.13.4
*/

export interface CreateTodoRequest {
	title: string;
}

export interface Todo {
	id: string;
	title: string;
	completed: boolean;
	createdAt: string;
	updatedAt: string;
}

export interface TodoListResponse {
	todos: Todo[];
	activeCount: number;
	completedCount: number;
}

export interface UpdateTodoRequest {
	title?: string;
	completed?: boolean;
}